[#9] 컴포넌트, 레이아웃 따로 분리
